feat: function exports and tests
