+++ lib/functions/summary/summary.test.ts
@@ -1,5 +1,6 @@
-import { QueryPlan } from "../types";
-import { buildSummary, Summary } from "./summary";
+import { QueryPlan } from "../../types";
+import { buildPieCharts, PieChartData } from "./buildPieChart";
+import { buildSummary, Summary } from "./buildSummary";
 
 const queryPlan: QueryPlan = {
   planInfo: {
@@ -124,7 +125,8 @@ const queryPlan: QueryPlan = {
 
 describe("buildSummary", () => {
   it("gives nothing when given nothing", () => {
-    const result = buildSummary({
+    // *** Summary ***
+    const summaryResult = buildSummary({
       planInfo: {
         pivotType: "",
         pivotId: "",
@@ -207,13 +209,117 @@ describe("buildSummary", () => {
         RangeSharingPrimitiveAggregatesRetrieval: 0,
       },
     };
-    expect(result).toEqual(emptySummary);
+
+    expect(summaryResult).toEqual(emptySummary);
+
+    // *** Pie Chart ***
+    const pieResult = buildPieCharts(
+      {
+        DatabaseRetrieval: 0,
+        PrimitiveAnalysisAggregationRetrieval: 0,
+        PartialPrimitiveAggregatesRetrieval: 0,
+        PrimitiveResultsMerger: 0,
+        RangeSharingPrimitiveAggregatesRetrieval: 0,
+      },
+      summaryResult,
+    );
+
+    const emptyPieChart: PieChartData = {
+      groupedPieDataElaspedTimings: [
+        {
+          fill: "#800080",
+          name: "Database",
+          value: 0,
+        },
+        {
+          fill: "#FFA500",
+          name: "Engine",
+          value: 0,
+        },
+        {
+          fill: "#0000FF",
+          name: "Network",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "Providers",
+          value: 0,
+        },
+      ],
+      groupedPieDataRetrievalsTypeCounts: [
+        {
+          fill: "#800080",
+          name: "Database",
+          value: 0,
+        },
+        {
+          fill: "#FFA500",
+          name: "Engine",
+          value: 0,
+        },
+        {
+          fill: "#0000FF",
+          name: "Network",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "Providers",
+          value: 0,
+        },
+      ],
+      pieDataElapsedTimings: [
+        {
+          fill: "#800080",
+          name: "DatabaseRetrieval",
+          value: 0,
+        },
+      ],
+      pieDataRetrievalsTypeCounts: [
+        {
+          fill: "#800080",
+          name: "DatabaseRetrieval",
+          value: 0,
+        },
+        {
+          fill: "#FFA500",
+          name: "PrimitiveAnalysisAggregationRetrieval",
+          value: 0,
+        },
+        {
+          fill: "#0000FF",
+          name: "PartialPrimitiveAggregatesRetrieval",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "PrimitiveResultsMerger",
+          value: 0,
+        },
+        {
+          fill: "#00FF00",
+          name: "RangeSharingPrimitiveAggregatesRetrieval",
+          value: 0,
+        },
+      ],
+      retrievalsColors: {
+        DatabaseRetrieval: "#800080",
+        PartialPrimitiveAggregatesRetrieval: "#0000FF",
+        PrimitiveAnalysisAggregationRetrieval: "#FFA500",
+        PrimitiveResultsMerger: "#FF0000",
+        RangeSharingPrimitiveAggregatesRetrieval: "#00FF00",
+      },
+    };
+
+    expect(pieResult).toEqual(emptyPieChart);
   });
 
   it("builds a summary from a query plan", () => {
-    const result = buildSummary(queryPlan);
+    // *** Summary ***
+    const summaryResult = buildSummary(queryPlan);
 
-    const expected: Summary = {
+    const expectedSummary: Summary = {
       aggregateRetrievalsElapsedTimings: {
         PrimitiveAnalysisAggregationRetrieval: 37,
       },
@@ -251,6 +357,113 @@ describe("buildSummary", () => {
       },
     };
 
-    expect(result).toEqual(expected);
+    expect(summaryResult).toEqual(expectedSummary);
+
+    // *** Pie Chart ***
+    const pieResult = buildPieCharts(
+      {
+        DatabaseRetrieval: 2,
+        PrimitiveAnalysisAggregationRetrieval: 1,
+        PartialPrimitiveAggregatesRetrieval: 0,
+        PrimitiveResultsMerger: 0,
+        RangeSharingPrimitiveAggregatesRetrieval: 0,
+      },
+      summaryResult,
+    );
+
+    const expectedPieChart: PieChartData = {
+      groupedPieDataElaspedTimings: [
+        {
+          fill: "#FFA500",
+          name: "Engine",
+          value: 37,
+        },
+        {
+          fill: "#800080",
+          name: "Database",
+          value: 14,
+        },
+        {
+          fill: "#0000FF",
+          name: "Network",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "Providers",
+          value: 0,
+        },
+      ],
+      groupedPieDataRetrievalsTypeCounts: [
+        {
+          fill: "#800080",
+          name: "Database",
+          value: 2,
+        },
+        {
+          fill: "#FFA500",
+          name: "Engine",
+          value: 1,
+        },
+        {
+          fill: "#0000FF",
+          name: "Network",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "Providers",
+          value: 0,
+        },
+      ],
+      pieDataElapsedTimings: [
+        {
+          fill: "#FFA500",
+          name: "PrimitiveAnalysisAggregationRetrieval",
+          value: 37,
+        },
+        {
+          fill: "#800080",
+          name: "DatabaseRetrieval",
+          value: 14,
+        },
+      ],
+      pieDataRetrievalsTypeCounts: [
+        {
+          fill: "#800080",
+          name: "DatabaseRetrieval",
+          value: 2,
+        },
+        {
+          fill: "#FFA500",
+          name: "PrimitiveAnalysisAggregationRetrieval",
+          value: 1,
+        },
+        {
+          fill: "#0000FF",
+          name: "PartialPrimitiveAggregatesRetrieval",
+          value: 0,
+        },
+        {
+          fill: "#FF0000",
+          name: "PrimitiveResultsMerger",
+          value: 0,
+        },
+        {
+          fill: "#00FF00",
+          name: "RangeSharingPrimitiveAggregatesRetrieval",
+          value: 0,
+        },
+      ],
+      retrievalsColors: {
+        DatabaseRetrieval: "#800080",
+        PartialPrimitiveAggregatesRetrieval: "#0000FF",
+        PrimitiveAnalysisAggregationRetrieval: "#FFA500",
+        PrimitiveResultsMerger: "#FF0000",
+        RangeSharingPrimitiveAggregatesRetrieval: "#00FF00",
+      },
+    };
+
+    expect(pieResult).toEqual(expectedPieChart);
   });
 });
